fix: skip duplicate dashboard when one is already running
Move isDashboardRunning() from index.ts to dashboard/server.ts as a
shared export. Add the same health-check guard in orchestrator.ts so
that approve/reject/resume commands reuse the existing dashboard
instead of spawning a second server on a different port.
